+++ src/main/java/amp/awec/command/operation/CommandCylinder.java
@@ -1,9 +1,8 @@
 package amp.awec.command.operation;
 
 import amp.awec.command.CommandPlayerData;
-import amp.awec.command.argtypes.ArgumentTypePattern;
+import amp.awec.command.argtypes.ArgumentTypeBlockPattern;
 import amp.awec.operation.CylinderOperation;
-import amp.awec.operation.SphereOperation;
 import amp.awec.operation.WorldChange;
 import amp.awec.pattern.BlockPattern;
 import amp.awec.permission.WorldEditPermissions;
@@ -26,7 +25,7 @@ public class CommandCylinder implements CommandManager.CommandRegistry {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal("/cyl")
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
-				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypePattern.normal())
+				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypeBlockPattern.pattern())
 					.then(ArgumentBuilderRequired.argument("radius", ArgumentTypeInteger.integer(1, 255))
 						.executes(context -> {
 							return handleCylinderCommand(context, 1);