+++ src/main/java/amp/awec/command/operation/CommandHollowCylinder.java
@@ -1,8 +1,7 @@
 package amp.awec.command.operation;
 
 import amp.awec.command.CommandPlayerData;
-import amp.awec.command.argtypes.ArgumentTypePattern;
-import amp.awec.operation.CylinderOperation;
+import amp.awec.command.argtypes.ArgumentTypeBlockPattern;
 import amp.awec.operation.HollowCylinderOperation;
 import amp.awec.operation.WorldChange;
 import amp.awec.pattern.BlockPattern;
@@ -26,7 +25,7 @@ public class CommandHollowCylinder implements CommandManager.CommandRegistry {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal("/hcyl")
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
-				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypePattern.normal())
+				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypeBlockPattern.pattern())
 					.then(ArgumentBuilderRequired.argument("radius", ArgumentTypeInteger.integer(1, 255))
 						.executes(context -> {
 							return handleCylinderCommand(context, 1);