+++ src/main/java/amp/awec/command/operation/CommandHollowSphere.java
@@ -1,7 +1,7 @@
 package amp.awec.command.operation;
 
 import amp.awec.command.CommandPlayerData;
-import amp.awec.command.argtypes.ArgumentTypePattern;
+import amp.awec.command.argtypes.ArgumentTypeBlockPattern;
 import amp.awec.operation.HollowSphereOperation;
 import amp.awec.operation.WorldChange;
 import amp.awec.pattern.BlockPattern;
@@ -24,7 +24,7 @@ public class CommandHollowSphere implements CommandManager.CommandRegistry {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal("/hsphere")
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
-				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypePattern.normal())
+				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypeBlockPattern.pattern())
 					.then(ArgumentBuilderRequired.argument("radius", ArgumentTypeInteger.integer(1, 255))
 						.executes(context -> {
 							CommandSource source = (CommandSource) context.getSource();