+++ src/main/java/amp/awec/command/operation/CommandSet.java
@@ -2,7 +2,7 @@ package amp.awec.command.operation;
 import amp.awec.command.CommandPlayerData;
 import amp.awec.operation.WorldChange;
 import amp.awec.operation.SetOperation;
-import amp.awec.command.argtypes.ArgumentTypePattern;
+import amp.awec.command.argtypes.ArgumentTypeBlockPattern;
 import amp.awec.permission.WorldEditPermissions;
 import amp.awec.pattern.BlockPattern;
 import amp.awec.util.MessageHelper;
@@ -20,7 +20,7 @@ public class CommandSet implements CommandManager.CommandRegistry {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal("/set")
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
-				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypePattern.normal())
+				.then(ArgumentBuilderRequired.argument("pattern", ArgumentTypeBlockPattern.pattern())
 					.executes(context -> {
 						CommandSource source = (CommandSource) context.getSource();
 						CommandPlayerData playerData = CommandPlayerData.get(source);