+++ src/main/java/amp/awec/pattern/BlockPatternException.java
@@ -1,7 +1,0 @@
-package amp.awec.pattern;
-
-public class BlockPatternException extends Exception {
-	public BlockPatternException(String errorMessage) {
-		super(errorMessage);
-	}
-}