+++ src/main/java/amp/awec/util/PatternType.java
@@ -1,6 +1,0 @@
-package amp.awec.util;
-
-public enum PatternType {
-	NORMAL,
-	REPLACE
-}